Added switching Scene

+++ src/main/java/at/ac/fhcampuswien/fhmdb/database/WatchlistRepository.java
@@ -1,10 +1,7 @@
 package at.ac.fhcampuswien.fhmdb.database;
 
 import at.ac.fhcampuswien.fhmdb.exceptions.DatabaseException;
-import at.ac.fhcampuswien.fhmdb.models.Genre;
 import com.j256.ormlite.dao.Dao;
-import javafx.scene.chart.PieChart;
-
 import java.sql.SQLException;
 import java.util.List;
 
@@ -17,12 +14,15 @@ public class WatchlistRepository {
 
     public void removeFromWatchlist(WatchlistMovieEntity movie) throws DatabaseException {
         try {
-            watchlistMovieDao.delete(movie);
+            if (getAll().contains(movie))
+            {
+                watchlistMovieDao.delete(movie);
+            }
         } catch (SQLException e) {
             throw new DatabaseException();
         }
-        //watchlistMovieDao.delete(movie);
     }
+
     public List<WatchlistMovieEntity> getAll() throws DatabaseException {
         try {
             return watchlistMovieDao.queryForAll();
@@ -30,8 +30,6 @@ public class WatchlistRepository {
             throw new DatabaseException();
         }
     }
-        //return watchlistMovieDao.queryForAll();
-
 
     public void addToWatchlist(WatchlistMovieEntity movie) throws DatabaseException {
         try {
@@ -44,8 +42,6 @@ public class WatchlistRepository {
         } catch (SQLException e) {
             throw new DatabaseException();
         }
-
-        //watchlistMovieDao.create(movie);
     }
 
 }
